Material classes fixes, preparing for recipes.

+++ diorite-api/src/main/java/org/diorite/material/items/BowMat.java
@@ -178,7 +178,7 @@ public class BowMat extends BasicToolMat
      */
     public static void register(final BowMat element)
     {
-        byID.put(element.getType(), element);
+        byID.put((short) element.getType(), element);
         byName.put(element.getTypeName(), element);
     }
