+++ diorite-api/src/main/java/org/diorite/material/items/RabbitHideMat.java
@@ -112,7 +112,7 @@ public class RabbitHideMat extends ItemMaterialData
      */
     public static void register(final RabbitHideMat element)
     {
-        byID.put(element.getType(), element);
+        byID.put((short) element.getType(), element);
         byName.put(element.getTypeName(), element);
     }
